+++ src/components/generic/index.ts
@@ -4,6 +4,7 @@ export * from './BackToTopBtn'
 export * from './Explanation'
 export * from './FancyHorizRule'
 export * from './HideOnScroll'
+export * from './MarkdownWithRouteLinks'
 export * from './ReadMore'
 export * from './ScrollToTopOnMount'
 export * from './ShareButtons'
